Refactored span context injection/extraction.
