+++ config
@@ -14,6 +14,8 @@ NGX_ADDON_DEPS="$NGX_ADDON_DEPS \
 NGX_ADDON_SRCS="$NGX_ADDON_SRCS \
   $ngx_addon_dir/src/ngx_http_opentracing_module.cpp \
   $ngx_addon_dir/src/ngx_script.cpp \
+  $ngx_addon_dir/src/extract_span_context.cpp \
+  $ngx_addon_dir/src/inject_span_context.cpp \
   $ngx_addon_dir/src/opentracing_request_processor.cpp"
 NGX_ADDON_DEPS="$NGX_ADDON_DEPS \
   $ngx_addon_dir/src/ngx_http_opentracing_conf.h \
