+++ src/opentracing_request_processor.cpp
@@ -12,6 +12,12 @@ extern ngx_module_t ngx_http_opentracing_module;
 }
 
 namespace ngx_opentracing {
+lightstep::SpanContext extract_span_context(lightstep::Tracer &tracer,
+                                            const ngx_http_request_t *request);
+
+void inject_span_context(lightstep::Tracer &tracer, ngx_http_request_t *request,
+                         const lightstep::SpanContext span_context);
+
 lightstep::Tracer make_tracer(const tracer_options_t &);
 
 static bool
@@ -24,123 +30,6 @@ is_opentracing_enabled(const ngx_http_request_t *request,
     // Only trace subrequests if `log_subrequest` is enabled; otherwise the
     // spans won't be finished.
     return loc_conf->enable && core_loc_conf->log_subrequest;
-}
-
-static bool insert_header(ngx_http_request_t *request, ngx_str_t key,
-                          ngx_str_t value) {
-  auto header = static_cast<ngx_table_elt_t *>(
-      ngx_list_push(&request->headers_in.headers));
-  if (!header)
-    return false;
-  header->hash = 1;
-  header->key = key;
-  header->lowcase_key = key.data;
-  header->value = value;
-  return true;
-}
-
-static bool set_headers(ngx_http_request_t *request,
-                        std::vector<std::pair<ngx_str_t, ngx_str_t>> &headers) {
-  if (headers.empty())
-    return true;
-
-  // If header keys are already in the request, overwrite the values instead of
-  // inserting a new header.
-  //
-  // It may be possible in some cases to use nginx's hashes to look up the
-  // entries faster, but then we'd have to handle the special case of when a
-  // header element isn't hashed yet. Iterating over the header entries all the
-  // time keeps things simple.
-  for_each<ngx_table_elt_t>(
-      request->headers_in.headers, [&](ngx_table_elt_t &header) {
-        auto i = std::find_if(
-            headers.begin(), headers.end(),
-            [&](const std::pair<ngx_str_t, ngx_str_t> &key_value) {
-              const auto &key = key_value.first;
-              return header.key.len == key.len &&
-                     ngx_strncmp(reinterpret_cast<char *>(header.lowcase_key),
-                                 reinterpret_cast<char *>(key.data),
-                                 key.len) == 0;
-
-            });
-        if (i == headers.end())
-          return;
-        header.value = i->second;
-        headers.erase(i);
-      });
-
-  // Any header left in `headers` doesn't already have a key in the request, so
-  // create a new entry for it.
-  for (const auto &key_value : headers) {
-    if (!insert_header(request, key_value.first, key_value.second)) {
-      ngx_log_error(NGX_LOG_ERR, request->connection->log, 0,
-                    "failed to insert header");
-      return false;
-    }
-  }
-  return true;
-}
-
-namespace {
-class NgxHeaderCarrierWriter : public lightstep::BasicCarrierWriter {
-public:
-  NgxHeaderCarrierWriter(ngx_http_request_t *request,
-                         std::vector<std::pair<ngx_str_t, ngx_str_t>> &headers,
-                         bool &was_successful)
-      : request_{request}, headers_{headers}, was_successful_{was_successful} {
-    was_successful_ = true;
-  }
-
-  void Set(const std::string &key, const std::string &value) const override {
-    if (!was_successful_)
-      return;
-    auto ngx_key = to_lower_ngx_str(request_->pool, key);
-    if (!ngx_key.data) {
-      ngx_log_error(NGX_LOG_ERR, request_->connection->log, 0,
-                    "failed to allocate header key");
-      was_successful_ = false;
-      return;
-    }
-    auto ngx_value = to_ngx_str(request_->pool, value);
-    if (!ngx_value.data) {
-      ngx_log_error(NGX_LOG_ERR, request_->connection->log, 0,
-                    "failed to allocate header value");
-      was_successful_ = false;
-      return;
-    }
-    headers_.emplace_back(ngx_key, ngx_value);
-  }
-
-private:
-  ngx_http_request_t *request_;
-  std::vector<std::pair<ngx_str_t, ngx_str_t>> &headers_;
-  bool &was_successful_;
-};
-}
-
-namespace {
-class NgxHeaderCarrierReader : public lightstep::BasicCarrierReader {
-public:
-  explicit NgxHeaderCarrierReader(const ngx_http_request_t *request)
-      : request_{request} {}
-
-  void ForeachKey(
-      std::function<void(const std::string &, const std::string &value)> f)
-      const {
-    std::string key, value;
-    for_each<ngx_table_elt_t>(
-        request_->headers_in.headers, [&](const ngx_table_elt_t &header) {
-          key.assign(reinterpret_cast<char *>(header.lowcase_key),
-                     header.key.len);
-          value.assign(reinterpret_cast<char *>(header.value.data),
-                       header.value.len);
-          f(key, value);
-        });
-  }
-
-private:
-  const ngx_http_request_t *request_;
-};
 }
 
 static void add_script_tag(ngx_http_request_t *request, lightstep::Span &span,
@@ -192,19 +81,7 @@ start_span(ngx_http_request_t *request,
                                 });
 
   // Inject the span's context into the request headers.
-  std::vector<std::pair<ngx_str_t, ngx_str_t>> headers;
-  bool was_successful = true;
-  auto carrier_writer =
-      NgxHeaderCarrierWriter{request, headers, was_successful};
-  was_successful =
-      tracer.Inject(span.context(), lightstep::CarrierFormat::HTTPHeaders,
-                    carrier_writer) &&
-      was_successful;
-  if (was_successful)
-    was_successful = set_headers(request, headers);
-  if (!was_successful)
-    ngx_log_error(NGX_LOG_ERR, request->connection->log, 0,
-                  "Tracer.inject() failed");
+  inject_span_context(tracer, request, span.context());
 
   return span;
 }
@@ -225,9 +102,7 @@ void OpenTracingRequestProcessor::before_response(ngx_http_request_t *request) {
       return;
     // No span has been created for this request yet. Check if there's any
     // parent span context in the request headers and start a new one.
-    auto carrier_reader = NgxHeaderCarrierReader{request};
-    auto parent_span_context =
-        tracer_.Extract(lightstep::CarrierFormat::HTTPHeaders, carrier_reader);
+    auto parent_span_context = extract_span_context(tracer_, request);
     auto span = start_span(request, core_loc_conf, loc_conf, tracer_,
                            parent_span_context, lightstep::ChildOfRef);
     active_spans_.emplace(request, std::move(span));
